Weighted contingency coefficients

Compute (weighted) pairwise contingency coefficients.

contingencyWt(x, ...)

Arguments

x

for the default method, a vector that can be interpreted as factor. For the matrix and data.frame methods, the columns should be interpretable as factors.

...

for the generic function, arguments to be passed down to the methods, otherwise ignored.

Value

For the default method, the (weighted) contingency coefficient of x and y.

For the matrix and data.frame method, a matrix of (weighted) pairwise contingency coefficients for all combinations of columns. Elements below the diagonal are NA.

Details

The function tableWt is used for the computation of the corresponding pairwise contingency tables. The following methods are implemented:

contingencyWt.default(x, y, weights = NULL, ...)

contingencyWt.matrix(x, weights = NULL, ...)

contingencyWt.data.frame(x, weights = NULL, ...)

Additional parameters are:

y: a vector that can be interpreted as factor (for the default method)

weights: an optional numeric vector containing sample weights

References

Kendall, M.G. and Stuart, A. (1967) The Advanced Theory of Statistics, Volume 2: Inference and Relationship. Charles Griffin & Co Ltd, London, 2nd edition.

See also

Author

Andreas Alfons and Stefan Kraft

Examples

data(eusilcS)

## default method
contingencyWt(eusilcS$pl030, eusilcS$pb220a, weights = eusilcS$rb050)
#> [1] 0.1749834

## data.frame method
basic <- c("age", "rb090", "hsize", "pl030", "pb220a")
contingencyWt(eusilcS[, basic], weights = eusilcS$rb050)
#> Warning: Chi-squared approximation may be incorrect
#> Warning: Chi-squared approximation may be incorrect
#> Warning: Chi-squared approximation may be incorrect
#> Warning: Chi-squared approximation may be incorrect
#> Warning: Chi-squared approximation may be incorrect
#> Warning: Chi-squared approximation may be incorrect
#>           rb090      hsize     pl030    pb220a
#> age   0.1454171 0.55819600 0.7203099 0.1751097
#> rb090        NA 0.07698323 0.3813192 0.0167801
#> hsize        NA         NA 0.3972100 0.1801206
#> pl030        NA         NA        NA 0.1749834