Synthetic EU-SILC 2013 survey data

This data set is synthetically generated from real Austrian EU-SILC (European Union Statistics on Income and Living Conditions) data 2013.

Format

A data frame with 13513 observations on the following 62 variables.

db030

integer; the household ID.

hsize

integer; the number of persons in the household.

db040

factor; the federal state in which the household is located (levels Burgenland, Carinthia, Lower Austria, Salzburg, Styria, Tyrol, Upper Austria, Vienna and Vorarlberg).

age

integer; the person's age.

rb090

factor; the person's gender (levels male and female).

pid

personal ID

weight

sampling weights

pl031

factor; the person's economic status (levels 1 = working full time, 2 = working part time, 3 = unemployed, 4 = pupil, student, further training or unpaid work experience or in compulsory military or community service, 5 = in retirement or early retirement or has given up business, 6 = permanently disabled or/and unfit to work or other inactive person, 7 = fulfilling domestic tasks and care responsibilities).

pb220a

factor; the person's citizenship (levels AT, EU and Other).

pb190

for details, see Eurostat's code book

pe040

for details, see Eurostat's code book

pl111

for details, see Eurostat's code book

pgrossIncomeCat

for details, see Eurostat's code book

pgrossIncome

for details, see Eurostat's code book

hgrossIncomeCat

for details, see Eurostat's code book

hgrossIncome

for details, see Eurostat's code book

hgrossminusCat

for details, see Eurostat's code book

hgrossminus

for details, see Eurostat's code book

py010g

for details, see Eurostat's code book

py021g

for details, see Eurostat's code book

py050g

for details, see Eurostat's code book

py080g

for details, see Eurostat's code book

py090g

for details, see Eurostat's code book

py100g

for details, see Eurostat's code book

py110g

for details, see Eurostat's code book

py120g

for details, see Eurostat's code book

py130g

for details, see Eurostat's code book

py140g

for details, see Eurostat's code book

hy040g

for details, see Eurostat's code book

hy050g

for details, see Eurostat's code book

hy060g

for details, see Eurostat's code book

hy070g

for details, see Eurostat's code book

hy080g

for details, see Eurostat's code book

hy090g

for details, see Eurostat's code book

hy100g

for details, see Eurostat's code book

hy110g

for details, see Eurostat's code book

hy120g

for details, see Eurostat's code book

hy130g

for details, see Eurostat's code book

hy140g

for details, see Eurostat's code book

rb250

for details, see Eurostat's code book

p119000

for details, see Eurostat's code book

p038003f

for details, see Eurostat's code book

p118000i

for details, see Eurostat's code book

aktivi

for details, see Eurostat's code book

erwintensneu

for details, see Eurostat's code book

rb050

for details, see Eurostat's code book

pb040

for details, see Eurostat's code book

hb030

for details, see Eurostat's code book

px030

for details, see Eurostat's code book

rx030

for details, see Eurostat's code book

pb030

for details, see Eurostat's code book

rb030

for details, see Eurostat's code book

hx040

for details, see Eurostat's code book

pb150

for details, see Eurostat's code book

rx020

for details, see Eurostat's code book

px020

for details, see Eurostat's code book

hx050

for details, see Eurostat's code book

eqInc

for details, see Eurostat's code book

hy010

for details, see Eurostat's code book

hy020

for details, see Eurostat's code book

hy022

for details, see Eurostat's code book

hy023

for details, see Eurostat's code book

Source

This is a synthetic data set based on Austrian EU-SILC data from 2013. The original sample was provided by Statistics Austria.

Details

The data set consists of 5977 households and is used as sample data in some of the examples in package simPop. Note that it is included for illustrative purposes only. The sample weights do not reflect the true population sizes of Austria and its regions.

62 variables of the original survey are simulated for this example data set. The variable names are rather cryptic codes, but these are the standardized names used by the statistical agencies. Furthermore, the variables hsize, age and netIncome are not included in the standardized format of EU-SILC data, but have been derived from other variables for convenience.

References

Eurostat (2013) Description of target variables: Cross-sectional and longitudinal.

Author

Matthias Templ

Examples

data(eusilc13puf)
str(eusilc13puf)
#> 'data.frame':	13513 obs. of  62 variables:
#>  $ db030          : int  30 53 168 296 317 337 438 576 747 799 ...
#>  $ hsize          : Factor w/ 8 levels "1","2","3","4",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ db040          : Factor w/ 9 levels "AT11","AT12",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ age            : Factor w/ 102 levels "-1","0","1","2",..: 93 55 59 93 31 68 38 38 72 81 ...
#>  $ rb090          : Factor w/ 2 levels "male","female": 2 1 2 2 1 1 1 1 2 1 ...
#>  $ pid            : chr  "30.1" "53.1" "168.1" "296.1" ...
#>  $ weight         : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ pl031          : Factor w/ 11 levels "1","2","3","4",..: NA 1 2 NA 1 1 1 1 1 1 ...
#>  $ pb220a         : Factor w/ 3 levels "AT","EU","Other": NA 1 1 NA 1 1 1 1 1 1 ...
#>  $ pb190          : Factor w/ 5 levels "1","2","3","4",..: NA 2 2 NA 2 1 2 1 1 1 ...
#>  $ pe040          : Factor w/ 7 levels "0","1","2","3",..: NA 4 4 NA 6 3 4 4 3 3 ...
#>  $ pl111          : Factor w/ 13 levels "a","b-e","f",..: NA 2 10 NA 2 3 9 9 3 3 ...
#>  $ pgrossIncomeCat: Factor w/ 22 levels "[-1.87e+03,-1.57e+03)",..: 11 19 18 11 20 11 20 19 15 11 ...
#>  $ pgrossIncome   : num  0 27307 22794 0 37470 ...
#>  $ hgrossIncomeCat: Factor w/ 11 levels "0","(0,91.8]",..: 2 5 4 2 3 10 3 3 10 2 ...
#>  $ hgrossIncome   : num  55.6 1448.7 416.1 62.8 128.9 ...
#>  $ hgrossminusCat : Factor w/ 21 levels "[-1.87e+03,-1.33e+03)",..: 18 16 19 20 18 16 17 17 16 15 ...
#>  $ hgrossminus    : num  16551 11893 23949 31326 16151 ...
#>  $ py010g         : num  0 25559 22794 0 NA ...
#>  $ py021g         : num  0 0 0 0 NA 0 0 0 0 0 ...
#>  $ py050g         : num  0 0 0 0 NA 0 0 0 0 0 ...
#>  $ py080g         : num  0 0 0 0 NA 0 0 0 0 0 ...
#>  $ py090g         : num  0 1748 0 0 NA ...
#>  $ py100g         : num  0 0 0 0 NA 0 0 0 0 0 ...
#>  $ py110g         : num  0 0 0 0 NA 0 0 0 0 0 ...
#>  $ py120g         : num  0 0 0 0 NA 0 0 0 0 0 ...
#>  $ py130g         : num  0 0 0 0 NA 0 0 0 0 0 ...
#>  $ py140g         : num  0 0 0 0 NA 0 0 0 0 0 ...
#>  $ hy040g         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ hy050g         : num  0 389 34 0 0 ...
#>  $ hy060g         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ hy070g         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ hy080g         : num  0 584 0 0 0 ...
#>  $ hy090g         : num  0 18.45 2.3 0 3.09 ...
#>  $ hy100g         : num  0 0 0 0 0 ...
#>  $ hy110g         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ hy120g         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ hy130g         : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ hy140g         : num  0 457 380 0 126 ...
#>  $ rb250          : Factor w/ 2 levels "11","14": NA NA NA NA NA NA NA NA NA NA ...
#>  $ p119000        : Factor w/ 51 levels "-1","8","10",..: NA NA NA NA NA NA NA NA NA NA ...
#>  $ p038003f       : Factor w/ 4 levels "-2","-1","1",..: 1 3 1 1 1 1 3 3 3 1 ...
#>  $ p118000i       : Factor w/ 11 levels "0","1","2","3",..: 5 3 6 5 11 3 11 11 5 2 ...
#>  $ aktivi         : Factor w/ 13 levels "0","1","2","3",..: 1 13 1 1 13 1 13 13 1 1 ...
#>  $ erwintensneu   : Factor w/ 4 levels "-1","1","2","3": 3 4 3 3 1 1 4 4 3 1 ...
#>  $ rb050          : num  576 576 576 576 576 ...
#>  $ pb040          : num  576 576 576 576 576 ...
#>  $ hb030          : int  30 53 168 296 317 337 438 576 747 799 ...
#>  $ px030          : int  30 53 168 296 317 337 438 576 747 799 ...
#>  $ rx030          : int  30 53 168 296 317 337 438 576 747 799 ...
#>  $ pb030          : chr  "30.1" "53.1" "168.1" "296.1" ...
#>  $ rb030          : chr  "30.1" "53.1" "168.1" "296.1" ...
#>  $ hx040          : Factor w/ 8 levels "1","2","3","4",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ pb150          : Factor w/ 2 levels "male","female": 2 1 2 2 1 1 1 1 2 1 ...
#>  $ rx020          : num  91 53 57 91 29 66 36 36 70 79 ...
#>  $ px020          : num  91 53 57 91 29 66 36 36 70 79 ...
#>  $ hx050          : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ eqInc          : num  0 27842 22451 0 -123 ...
#>  $ hy010          : num  0 28299 22830 0 NA ...
#>  $ hy020          : num  0 27842 22451 0 NA ...
#>  $ hy022          : num  0 25705 22417 0 NA ...
#>  $ hy023          : num  0 25705 22417 0 NA ...