Synthetic EU-SILC data

This data set is synthetically generated from real Austrian EU-SILC (European Union Statistics on Income and Living Conditions) data.

Format

A data.frame with 58 654 observations on the following 28 variables:

hid

integer; the household ID.

region

factor; the federal state in which the household is located (levels Burgenland, Carinthia, Lower Austria, Salzburg, Styria, Tyrol, Upper Austria, Vienna and Vorarlberg).

hsize

integer; the number of persons in the household.

eqsize

numeric; the equivalized household size according to the modified OECD scale.

eqIncome

numeric; a simplified version of the equivalized household income.

pid

integer; the personal ID.

id

the household ID combined with the personal ID. The first five digits represent the household ID, the last two digits the personal ID (both with leading zeros).

age

integer; the person's age.

gender

factor; the person's gender (levels male and female).

ecoStat

factor; the person's economic status (levels 1 = working full time, 2 = working part time, 3 = unemployed, 4 = pupil, student, further training or unpaid work experience or in compulsory military or community service, 5 = in retirement or early retirement or has given up business, 6 = permanently disabled or/and unfit to work or other inactive person, 7 = fulfilling domestic tasks and care responsibilities).

citizenship

factor; the person's citizenship (levels AT, EU and Other).

py010n

numeric; employee cash or near cash income (net).

py050n

numeric; cash benefits or losses from self-employment (net).

py090n

numeric; unemployment benefits (net).

py100n

numeric; old-age benefits (net).

py110n

numeric; survivor's benefits (net).

py120n

numeric; sickness benefits (net).

py130n

numeric; disability benefits (net).

py140n

numeric; education-related allowances (net).

hy040n

numeric; income from rental of a property or land (net).

hy050n

numeric; family/children related allowances (net).

hy070n

numeric; housing allowances (net).

hy080n

numeric; regular inter-household cash transfer received (net).

hy090n

numeric; interest, dividends, profit from capital investments in unincorporated business (net).

hy110n

numeric; income received by people aged under 16 (net).

hy130n

numeric; regular inter-household cash transfer paid (net).

hy145n

numeric; repayments/receipts for tax adjustment (net).

main

logical; indicates the main income holder (i.e., the person with the highest income) of each household.

Source

This is a synthetic data set based on Austrian EU-SILC data from 2006. The original sample was provided by Statistics Austria.

Details

The data set is used as population data in some of the examples in package simFrame. Note that it is included for illustrative purposes only. It consists of 25 000 households, hence it does not represent the true population sizes of Austria and its regions.

Only a few of the large number of variables in the original survey are included in this example data set. Some variable names are different from the standardized names used by the statistical agencies, as the latter are rather cryptic codes. Furthermore, the variables hsize, eqsize, eqIncome and age are not included in the standardized format of EU-SILC data, but have been derived from other variables for convenience. Moreover, some very sparse income components were not included in the the generation of this synthetic data set. Thus the equivalized household income is computed from the available income components.

References

Eurostat (2004) Description of target variables: Cross-sectional and longitudinal. EU-SILC 065/04, Eurostat.

Examples

data(eusilcP)
summary(eusilcP)
#>       hid                  region          hsize           eqsize     
#>  Min.   :    1   Vienna       :11657   2      :14128   Min.   :1.000  
#>  1st Qu.: 6262   Lower Austria:11127   4      :13180   1st Qu.:1.500  
#>  Median :12465   Upper Austria:10310   3      :12429   Median :2.000  
#>  Mean   :12488   Styria       : 8142   1      : 8602   Mean   :1.943  
#>  3rd Qu.:18719   Tyrol        : 4796   5      : 6745   3rd Qu.:2.400  
#>  Max.   :25000   Carinthia    : 4111   6      : 2094   Max.   :4.500  
#>                  (Other)      : 8511   (Other): 1476                  
#>     eqIncome           pid            id                 age       
#>  Min.   :     0   Min.   :1.00   Length:58654       Min.   :-1.00  
#>  1st Qu.: 13539   1st Qu.:1.00   Class :character   1st Qu.:22.00  
#>  Median : 18322   Median :2.00   Mode  :character   Median :40.00  
#>  Mean   : 20163   Mean   :2.07                      Mean   :39.75  
#>  3rd Qu.: 24277   3rd Qu.:3.00                      3rd Qu.:57.00  
#>  Max.   :179946   Max.   :9.00                      Max.   :97.00  
#>                                                                    
#>     gender         ecoStat      citizenship       py010n           py050n      
#>  male  :28539   1      :20900   AT   :44066   Min.   :     0   Min.   : -6895  
#>  female:30115   5      :12836   EU   : 1257   1st Qu.:     0   1st Qu.:     0  
#>                 7      : 4607   Other: 3162   Median :  2382   Median :     0  
#>                 2      : 4362   NA's :10169   Mean   :  9062   Mean   :  1288  
#>                 4      : 2921                 3rd Qu.: 16820   3rd Qu.:     0  
#>                 (Other): 2859                 Max.   :199075   Max.   :129874  
#>                 NA's   :10169                 NA's   :10169    NA's   :10169   
#>      py090n            py100n           py110n            py120n        
#>  Min.   :    0.0   Min.   :     0   Min.   :    0.0   Min.   :    0.00  
#>  1st Qu.:    0.0   1st Qu.:     0   1st Qu.:    0.0   1st Qu.:    0.00  
#>  Median :    0.0   Median :     0   Median :    0.0   Median :    0.00  
#>  Mean   :  444.6   Mean   :  3713   Mean   :   72.9   Mean   :   51.22  
#>  3rd Qu.:    0.0   3rd Qu.:     0   3rd Qu.:    0.0   3rd Qu.:    0.00  
#>  Max.   :29887.1   Max.   :101777   Max.   :22546.8   Max.   :46398.44  
#>  NA's   :10169     NA's   :10169    NA's   :10169     NA's   :10169     
#>      py130n            py140n             hy040n             hy050n      
#>  Min.   :    0.0   Min.   :    0.00   Min.   : -2962.5   Min.   :-11857  
#>  1st Qu.:    0.0   1st Qu.:    0.00   1st Qu.:     0.0   1st Qu.:     0  
#>  Median :    0.0   Median :    0.00   Median :     0.0   Median :     0  
#>  Mean   :  393.7   Mean   :   41.73   Mean   :   879.9   Mean   :  2826  
#>  3rd Qu.:    0.0   3rd Qu.:    0.00   3rd Qu.:     0.0   3rd Qu.:  4558  
#>  Max.   :53183.6   Max.   :18643.46   Max.   :129586.6   Max.   :118309  
#>  NA's   :10169     NA's   :10169                                         
#>      hy070n             hy080n             hy090n              hy110n        
#>  Min.   :    0.00   Min.   :     0.0   Min.   :  -457.46   Min.   :    0.00  
#>  1st Qu.:    0.00   1st Qu.:     0.0   1st Qu.:     0.75   1st Qu.:    0.00  
#>  Median :    0.00   Median :     0.0   Median :    58.45   Median :    0.00  
#>  Mean   :   93.12   Mean   :   744.6   Mean   :   462.45   Mean   :   32.97  
#>  3rd Qu.:    0.00   3rd Qu.:     0.0   3rd Qu.:   234.78   3rd Qu.:    0.00  
#>  Max.   :17954.97   Max.   :124206.2   Max.   :112011.03   Max.   :14506.49  
#>                                                                              
#>      hy130n            hy145n            main        
#>  Min.   :-5489.6   Min.   :-29519.3   Mode :logical  
#>  1st Qu.:    0.0   1st Qu.:  -256.8   FALSE:33654    
#>  Median :    0.0   Median :     0.0   TRUE :25000    
#>  Mean   :  339.1   Mean   :  -108.8                  
#>  3rd Qu.:    0.0   3rd Qu.:     0.0                  
#>  Max.   :40762.9   Max.   : 49768.0                  
#>