Synthetic EU-SILC survey data

This data set is synthetically generated from real Austrian EU-SILC (European Union Statistics on Income and Living Conditions) data.

Format

A data frame with 11725 observations on the following 18 variables.

db030

integer; the household ID.

hsize

integer; the number of persons in the household.

db040

factor; the federal state in which the household is located (levels Burgenland, Carinthia, Lower Austria, Salzburg, Styria, Tyrol, Upper Austria, Vienna and Vorarlberg).

age

integer; the person's age.

rb090

factor; the person's gender (levels male and female).

pl030

factor; the person's economic status (levels 1 = working full time, 2 = working part time, 3 = unemployed, 4 = pupil, student, further training or unpaid work experience or in compulsory military or community service, 5 = in retirement or early retirement or has given up business, 6 = permanently disabled or/and unfit to work or other inactive person, 7 = fulfilling domestic tasks and care responsibilities).

pb220a

factor; the person's citizenship (levels AT, EU and Other).

netIncome

numeric; the personal net income.

py010n

numeric; employee cash or near cash income (net).

py050n

numeric; cash benefits or losses from self-employment (net).

py090n

numeric; unemployment benefits (net).

py100n

numeric; old-age benefits (net).

py110n

numeric; survivor's benefits (net).

py120n

numeric; sickness benefits (net).

py130n

numeric; disability benefits (net).

py140n

numeric; education-related allowances (net).

db090

numeric; the household sample weights.

rb050

numeric; the personal sample weights.

Source

This is a synthetic data set based on Austrian EU-SILC data from 2006. The original sample was provided by Statistics Austria.

Details

The data set consists of 4641 households and is used as sample data in some of the examples in package simPopulation. Note that it is included for illustrative purposes only. The sample weights do not reflect the true population sizes of Austria and its regions. The resulting population data is about 100 times smaller than the real population size to save computation time.

Only a few of the large number of variables in the original survey are included in this example data set. The variable names are rather cryptic codes, but these are the standardized names used by the statistical agencies. Furthermore, the variables hsize, age and netIncome are not included in the standardized format of EU-SILC data, but have been derived from other variables for convenience.

References

Eurostat (2004) Description of target variables: Cross-sectional and longitudinal. EU-SILC 065/04, Eurostat.

Examples

data(eusilcS)
summary(eusilcS)
#>      db030          hsize                 db040           age       
#>  Min.   :   1   Min.   :1.000   Upper Austria:2120   Min.   :-1.00  
#>  1st Qu.:1152   1st Qu.:2.000   Lower Austria:2080   1st Qu.:20.00  
#>  Median :2339   Median :3.000   Styria       :1859   Median :39.00  
#>  Mean   :2327   Mean   :3.321   Vienna       :1789   Mean   :38.89  
#>  3rd Qu.:3495   3rd Qu.:4.000   Tyrol        :1119   3rd Qu.:55.00  
#>  Max.   :4641   Max.   :9.000   Carinthia    : 927   Max.   :96.00  
#>                                 (Other)      :1831                  
#>     rb090          pl030        pb220a       netIncome          py010n      
#>  male  :5764   1      :4207   AT   :8687   Min.   : -5809   Min.   :     0  
#>  female:5961   5      :2352   EU   : 210   1st Qu.:  6332   1st Qu.:     0  
#>                7      : 998   Other: 625   Median : 14229   Median :  2937  
#>                2      : 876   NA's :2203   Mean   : 14891   Mean   :  9272  
#>                4      : 604                3rd Qu.: 20966   3rd Qu.: 17040  
#>                (Other): 485                Max.   :185456   Max.   :185456  
#>                NA's   :2203                NA's   :2203     NA's   :2203    
#>      py050n           py090n            py100n           py110n        
#>  Min.   :-11937   Min.   :    0.0   Min.   :     0   Min.   :    0.00  
#>  1st Qu.:     0   1st Qu.:    0.0   1st Qu.:     0   1st Qu.:    0.00  
#>  Median :     0   Median :    0.0   Median :     0   Median :    0.00  
#>  Mean   :  1183   Mean   :  434.6   Mean   :  3525   Mean   :   65.84  
#>  3rd Qu.:     0   3rd Qu.:    0.0   3rd Qu.:     0   3rd Qu.:    0.00  
#>  Max.   : 97324   Max.   :28908.0   Max.   :101916   Max.   :16782.20  
#>  NA's   :2203     NA's   :2203      NA's   :2203     NA's   :2203      
#>      py120n             py130n          py140n             db090       
#>  Min.   :    0.00   Min.   :    0   Min.   :    0.00   Min.   : 3.575  
#>  1st Qu.:    0.00   1st Qu.:    0   1st Qu.:    0.00   1st Qu.: 5.998  
#>  Median :    0.00   Median :    0   Median :    0.00   Median : 6.914  
#>  Mean   :   41.34   Mean   :  330   Mean   :   38.66   Mean   : 6.978  
#>  3rd Qu.:    0.00   3rd Qu.:    0   3rd Qu.:    0.00   3rd Qu.: 7.823  
#>  Max.   :31425.57   Max.   :35467   Max.   :16553.36   Max.   :12.526  
#>  NA's   :2203       NA's   :2203    NA's   :2203                       
#>      rb050       
#>  Min.   : 3.575  
#>  1st Qu.: 5.998  
#>  Median : 6.914  
#>  Mean   : 6.978  
#>  3rd Qu.: 7.823  
#>  Max.   :12.526  
#>