Categorize (semi-)continuous variables

Categorize continuous or semi-continuous variables. This is a utility function that is useful for writing custom wrapper functions such as simEUSILC.

getCat(x, breaks, zeros = TRUE, right = FALSE)

Arguments

x

a numeric vector to be categorized.

breaks

a numeric vector of two or more break points.

zeros

a logical indicating whether x is semi-continuous, i.e., contains a considerable amount of zeros. See “Details” on how this affects the behavior of the function.

right

logical; if zeros is not TRUE, this indicates whether the intervals should be closed on the right (and open on the left) or vice versa.

Value

A factor containing the categories.

Details

If zeros is TRUE, 0 is added to the break points and treated as its own factor level. Consequently, intervals for negative values are left-closed and right-open, whereas intervals for positive values are left-open and right-closed.

See also

Author

Andreas Alfons

Examples

data(eusilcS)

## semi-continuous variable
breaks <- getBreaks(eusilcS$netIncome, 
    weights=eusilcS$rb050, equidist = FALSE)
netIncomeCat <- getCat(eusilcS$netIncome, breaks)
summary(netIncomeCat)
#> [-5.81e+03,-4.37e+03) [-4.37e+03,-3.11e+03)         [-3.11e+03,0) 
#>                     1                     1                     1 
#>                     0               (0,636]        (636,2.64e+03] 
#>                  1193                    83                   348 
#>   (2.64e+03,4.99e+03]   (4.99e+03,8.48e+03]   (8.48e+03,1.38e+04] 
#>                   437                   849                  1659 
#>   (1.38e+04,1.79e+04]   (1.79e+04,2.36e+04]   (2.36e+04,2.92e+04] 
#>                  1660                  1657                   814 
#>   (2.92e+04,3.57e+04]   (3.57e+04,5.62e+04]   (5.62e+04,1.85e+05] 
#>                   411                   329                    79 
#>                  NA's 
#>                  2203