Extract and modify variables from population or sample data stored in an object of class simPopObj-class.

Using samp samp<- it is possible to extract or rather modify variables of the sample data within slot data in slot sample of the simPopObj-class-object. Using pop pop<- it is possible to extract or rather modify variables of the synthetic population within in slot data in slot sample of the simPopObj-class-object.

Arguments

obj

An object of class simPopObj-class

var

variable name or index for the variable in slot 'samp' of object with the slot name to be accessed. If NULL, the entire dataset (sample or population) is returned.

value

Content replacing whatever the variable in slot var in obj currently holds.

Value

Returns an object of class simPopObj-class with the appropriate replacement.

See also

Author

Bernhard Meindl

Examples

data(eusilcS)
inp <- specifyInput(data=eusilcS, hhid="db030", hhsize="hsize", strata="db040",
weight="db090")
simPopObj <- simStructure(data=inp, method="direct", basicHHvars=c("age", "rb090"))

## get/set variables in sample-object of simPopObj
head(samp(simPopObj, var="age"))
#> [1] 72 66 56 67 70 46
samp(simPopObj, var="newVar") <- 1
head(samp(simPopObj, var="newVar"))
#> [1] 1 1 1 1 1 1
## deleting is also possible
samp(simPopObj, var="newvar") <- NULL
head(samp(simPopObj, var="newvar"))
#> Warning: No specified variable was found in the sample! Check input 'var'!
#> NULL
## extract multiple variables
head(samp(simPopObj, var=c("db030","db040")))
#>    db030         db040
#> 1:     1      Salzburg
#> 2:     1      Salzburg
#> 3:     2 Upper Austria
#> 4:     3        Styria
#> 5:     4 Upper Austria
#> 6:     4 Upper Austria

## get/set variables in pop-object of simPopObj
head(pop(simPopObj, var="age"))
#> [1] 66 53 68 85 39 74
pop(simPopObj, var="newVar") <- 1
head(pop(simPopObj, var="newVar"))
#> [1] 1 1 1 1 1 1
## deleting is also possible
pop(simPopObj, var="newvar") <- NULL
head(pop(simPopObj, var="newvar"))
#> Warning: No specified variable was found in the synthetic population! Check input 'var'!
#> NULL
## extract multiple variables
head(pop(simPopObj, var=c("db030","db040")))
#>    db030      db040
#> 1:     1 Burgenland
#> 2:     2 Burgenland
#> 3:     3 Burgenland
#> 4:     4 Burgenland
#> 5:     5 Burgenland
#> 6:     6 Burgenland