Synthetic GLSS survey data

This data set is synthetically generated from real GLSS (Ghana Living Standards Survey) data.

Format

A data frame with 36970 observations on the following 14 variables.

hhid

integer; the household ID.

hsize

integer; the number of persons in the household.

region

factor; the region in which the household is located (levels western, central, greater accra, volta, eastern, ashanti, brong ahafo, northern, upper east and upper west).

clust

factor; the enumeration area.

age

integer; the person's age.

sex

factor; the person's sex (levels male and female).

relate

factor; the relationship with the household head (levels head, spouse, child, grandchild, parent/parentlaw, son/daughterlaw, other relative, adopted child, househelp and non_relative).

nation

factor; the person's nationality (levels ghanaian birth, ghanaian naturalise, burkinabe, malian, nigerian, ivorian, togolese, liberian, other ecowas, other africa and other).

ethnic

factor; the person's ethnicity (levels akan, all other tribes, ewe, ga-dangbe, grusi, guan, gurma, mande and mole-dagbani).

religion

factor; the person's religion (levels catholic, anglican, presbyterian, methodist, pentecostal, spiritualist, other christian, moslem, traditional, no religion and other).

highest_degree

factor; the person's highest degree of education (levels none, mlsc, bece, voc/comm, teacher trng a, teacher trng b, gce 'o' level, ssce, gce 'a' level, tech/prof cert, tech/prof dip, hnd, bachelor, masters, doctorate and other).

occupation

factor; the person's occupation (levels armed forces and other security personnel, clerks, craft and related trades workers, elementary occupations, legislators, senior officials and managers, none, plant and machine operators and assemblers, professionals, service workers and shop and market sales workers, skilled agricultural and fishery workers, and technicians and associate professionals).

income

numeric; the person's annual income.

weight

numeric; the sample weights.

Source

This is a synthetic data set based on GLSS data from 2006. The original sample was provided by Ghana Statistical Service.

Details

The data set consists of 8700 households and is used as sample data in some of the examples in package simPopulation. Note that it is included for illustrative purposes only. The sample weights do not reflect the true population sizes of Ghana and its regions. The resulting population data is about 100 times smaller than the real population size to save computation time.

Only some of the variables in the original survey are included in this example data set. Furthermore, categories are aggregated for certain variables due to the large number of possible outcomes in the original survey data.

References

Ghana Statistical Service (2008) Ghana Living Standards Survey: Report of the fifth round.

Examples

data(ghanaS)
summary(ghanaS)
#>       hhid          hsize                  region          clust      
#>  Min.   :   1   Min.   : 1.000   ashanti      : 6316   5550   :  161  
#>  1st Qu.:2207   1st Qu.: 4.000   northern     : 4407   5468   :  157  
#>  Median :4381   Median : 6.000   greater accra: 4291   5535   :  153  
#>  Mean   :4368   Mean   : 6.153   eastern      : 3537   5500   :  143  
#>  3rd Qu.:6551   3rd Qu.: 8.000   brong ahafo  : 3275   5469   :  136  
#>  Max.   :8700   Max.   :23.000   western      : 3264   5441   :  133  
#>                                  (Other)      :11880   (Other):36087  
#>       age            sex                     relate     
#>  Min.   : 0.00   male  :18082   child           :17442  
#>  1st Qu.: 9.00   female:18888   head            : 8700  
#>  Median :19.00                  spouse          : 5289  
#>  Mean   :24.19                  other relative  : 2192  
#>  3rd Qu.:36.00                  grandchild      : 2157  
#>  Max.   :99.00                  parent/parentlaw:  464  
#>                                 (Other)         :  726  
#>                  nation                   ethnic                 religion   
#>  ghanaian birth     :36232   akan            :15102   pentecostal    :7072  
#>  ghanaian naturalise:  194   mole-dagbani    : 7122   moslem         :6596  
#>  togolese           :  177   ewe             : 4537   catholic       :5388  
#>  burkinabe          :   96   gurma           : 2889   other christian:5220  
#>  other ecowas       :   79   ga-dangbe       : 2563   methodist      :4062  
#>  other african      :   48   all other tribes: 1574   presbyterian   :2848  
#>  (Other)            :  144   (Other)         : 3183   (Other)        :5784  
#>         highest_degree 
#>  none          :28189  
#>  bece          : 3250  
#>  mslc          : 2696  
#>  ssce          :  980  
#>  hnd           :  350  
#>  tech/prof cert:  279  
#>  (Other)       : 1226  
#>                                              occupation        income        
#>  none                                             :20857   Min.   :     0.0  
#>  skilled agricultural and fishery workers         : 9315   1st Qu.:     0.0  
#>  craft and related trades workers                 : 2029   Median :     0.0  
#>  service workers and shop and market sales workers: 1877   Mean   :   329.2  
#>  elementary occupations                           : 1133   3rd Qu.:   194.8  
#>  technicians and associate professionals          :  502   Max.   :488701.7  
#>  (Other)                                          : 1257                     
#>      weight      
#>  Min.   : 0.748  
#>  1st Qu.: 4.238  
#>  Median : 5.901  
#>  Mean   : 6.125  
#>  3rd Qu.: 7.583  
#>  Max.   :23.183  
#>