Iterative Proportional Updating

Adjust sampling weights to given totals based on household-level and/or individual level constraints.

ipu2(
  dat,
  hid = NULL,
  conP = NULL,
  conH = NULL,
  epsP = 1e-06,
  epsH = 0.01,
  verbose = FALSE,
  w = NULL,
  bound = 4,
  maxIter = 200,
  meanHH = TRUE,
  allPthenH = TRUE,
  returnNA = TRUE,
  looseH = FALSE,
  numericalWeighting = NULL,
  check_hh_vars = TRUE,
  conversion_messages = FALSE
)

Arguments

dat

a data.table containing household ids (optionally), base weights (optionally), household and/or personal level variables (numerical or categorical) that should be fitted.

hid

name of the column containing the household-ids within dat or NULL if such a variable does not exist.

conP

list or (partly) named list defining the constraints on person level. The list elements are contingency tables in array representation with dimnames corresponding to the names of the relevant calibration variables in dat. If a numerical variable is to be calibrated, the respective list element has to be named with the name of that numerical variable. Otherwise the list element shoud NOT be named.

conH

list or (partly) named list defining the constraints on household level. The list elements are contingency tables in array representation with dimnames corresponding to the names of the relevant calibration variables in dat. If a numerical variable is to be calibrated, the respective list element has to be named with the name of that numerical variable. Otherwise the list element shoud NOT be named.

epsP

numeric value or list (of numeric values and/or arrays) specifying the convergence limit(s) for conP. The list can contain numeric values and/or arrays which must appear in the same order as the corresponding constraints in conP. Also, an array must have the same dimensions and dimnames as the corresponding constraint in conP.

epsH

numeric value or list (of numeric values and/or arrays) specifying the convergence limit(s) for conH. The list can contain numeric values and/or arrays which must appear in the same order as the corresponding constraints in conH. Also, an array must have the same dimensions and dimnames as the corresponding constraint in conH.

verbose

if TRUE, some progress information will be printed.

w

name if the column containing the base weights within dat or NULL if such a variable does not exist. In the latter case, every observation in dat is assigned a starting weight of 1.

bound

numeric value specifying the multiplier for determining the weight trimming boundary if the change of the base weights should be restricted, i.e. if the weights should stay between 1/bound*w and bound*w.

maxIter

numeric value specifying the maximum number of iterations that should be performed.

meanHH

if TRUE, every person in a household is assigned the mean of the person weights corresponding to the household. If "geometric", the geometric mean is used rather than the arithmetic mean.

allPthenH

if TRUE, all the person level calibration steps are performed before the houshold level calibration steps (and meanHH, if specified). If FALSE, the houshold level calibration steps (and meanHH, if specified) are performed after everey person level calibration step. This can lead to better convergence properties in certain cases but also means that the total number of calibration steps is increased.

returnNA

if TRUE, the calibrated weight will be set to NA in case of no convergence.

looseH

if FALSE, the actual constraints conH are used for calibrating all the hh weights. If TRUE, only the weights for which the lower and upper thresholds defined by conH and epsH are exceeded are calibrated. They are however not calibrated against the actual constraints conH but against these lower and upper thresholds, i.e. conH-conH*epsH and conH+conH*epsH.

numericalWeighting

If NULL computeLinear from the pacakge survey sd will be used.

check_hh_vars

If TRUE check for non-unique values inside of a household for variables in household constraints

conversion_messages

show a message, if inputs need to be reformatted. This can be useful for speed optimizations if ipu2 is called several times with similar inputs (for example bootstrapping)

Value

The function will return the input data dat with the calibrated weights calibWeight as an additional column as well as attributes. If no convergence has been reached in maxIter

steps, and returnNA is TRUE (the default), the column calibWeights will only consist of NAs. The attributes of the table are attributes derived from the data.table class as well as the following.

converged	Did the algorithm converge in maxIter steps?

iterations	The number of iterations performed.

conP, conH, epsP, epsH	See Arguments.

conP_adj, conH_adj	Adjusted versions of conP and conH

formP, formH	Formulas that were used to calculate conP_adj and conH_adj based on the output table.

Details

This function implements the weighting procedure described here.

conP and conH are contingency tables, which can be created with xtabs. The dimnames of those tables should match the names and levels of the corresponding columns in dat.

maxIter, epsP and epsH are the stopping criteria. epsP and epsH describe relative tolerances in the sense that \deqn{1-epsP < \frac{w_{i+1}}{w_i} < 1+epsP}{1-epsP < w(i+1)/w(i) < 1+epsP} will be used as convergence criterium. Here i is the iteration step and wi is the weight of a specific person at step i.

The algorithm performs best if all varables occuring in the constraints (conP and conH) as well as the household variable are coded as factor-columns in dat. Otherwise, conversions will be necessary which can be monitored with the conversion_messages argument. Setting check_hh_vars to FALSE can also incease the performance of the scheme.

See also

Author

Alexander Kowarik, Gregor de Cillia

Examples

library(data.table)
data(eusilcS)
setDT(eusilcS)
eusilcS <- eusilcS[, list(db030,hsize,db040,age,rb090,netIncome,db090,rb050)]

## rename columns
setnames(eusilcS, "rb090", "gender")
setnames(eusilcS, "db040", "state")
setnames(eusilcS, "db030", "household")
setnames(eusilcS, "rb050", "weight")

## some recoding
# generate age groups
eusilcS[, agegroup := cut(age, c(-Inf, 10*1:9, Inf), right = FALSE)]
#>        household hsize         state age gender netIncome    db090   weight
#>     1:         1     2      Salzburg  72   male  22675.48 7.822929 7.822929
#>     2:         1     2      Salzburg  66 female  16999.29 7.822929 7.822929
#>     3:         2     1 Upper Austria  56 female  19274.21 8.788089 8.788089
#>     4:         3     1        Styria  67 female  13319.13 8.108452 8.108452
#>     5:         4     3 Upper Austria  70 female  14365.57 7.509383 7.509383
#>    ---                                                                     
#> 11721:      4639     2     Carinthia  60 female  11405.85 6.863162 6.863162
#> 11722:      4640     2     Carinthia  60   male  15476.11 6.863162 6.863162
#> 11723:      4640     2     Carinthia  55 female      0.00 6.863162 6.863162
#> 11724:      4641     2        Vienna  56   male  30367.92 8.772773 8.772773
#> 11725:      4641     2        Vienna  52 female   6368.93 8.772773 8.772773
#>        agegroup
#>     1:  [70,80)
#>     2:  [60,70)
#>     3:  [50,60)
#>     4:  [60,70)
#>     5:  [70,80)
#>    ---         
#> 11721:  [60,70)
#> 11722:  [60,70)
#> 11723:  [50,60)
#> 11724:  [50,60)
#> 11725:  [50,60)
# some recoding of netIncome for reasons of simplicity
eusilcS[is.na(netIncome), netIncome := 0] 
#>        household hsize         state age gender netIncome    db090   weight
#>     1:         1     2      Salzburg  72   male  22675.48 7.822929 7.822929
#>     2:         1     2      Salzburg  66 female  16999.29 7.822929 7.822929
#>     3:         2     1 Upper Austria  56 female  19274.21 8.788089 8.788089
#>     4:         3     1        Styria  67 female  13319.13 8.108452 8.108452
#>     5:         4     3 Upper Austria  70 female  14365.57 7.509383 7.509383
#>    ---                                                                     
#> 11721:      4639     2     Carinthia  60 female  11405.85 6.863162 6.863162
#> 11722:      4640     2     Carinthia  60   male  15476.11 6.863162 6.863162
#> 11723:      4640     2     Carinthia  55 female      0.00 6.863162 6.863162
#> 11724:      4641     2        Vienna  56   male  30367.92 8.772773 8.772773
#> 11725:      4641     2        Vienna  52 female   6368.93 8.772773 8.772773
#>        agegroup
#>     1:  [70,80)
#>     2:  [60,70)
#>     3:  [50,60)
#>     4:  [60,70)
#>     5:  [70,80)
#>    ---         
#> 11721:  [60,70)
#> 11722:  [60,70)
#> 11723:  [50,60)
#> 11724:  [50,60)
#> 11725:  [50,60)
eusilcS[netIncome < 0, netIncome := 0] 
#>        household hsize         state age gender netIncome    db090   weight
#>     1:         1     2      Salzburg  72   male  22675.48 7.822929 7.822929
#>     2:         1     2      Salzburg  66 female  16999.29 7.822929 7.822929
#>     3:         2     1 Upper Austria  56 female  19274.21 8.788089 8.788089
#>     4:         3     1        Styria  67 female  13319.13 8.108452 8.108452
#>     5:         4     3 Upper Austria  70 female  14365.57 7.509383 7.509383
#>    ---                                                                     
#> 11721:      4639     2     Carinthia  60 female  11405.85 6.863162 6.863162
#> 11722:      4640     2     Carinthia  60   male  15476.11 6.863162 6.863162
#> 11723:      4640     2     Carinthia  55 female      0.00 6.863162 6.863162
#> 11724:      4641     2        Vienna  56   male  30367.92 8.772773 8.772773
#> 11725:      4641     2        Vienna  52 female   6368.93 8.772773 8.772773
#>        agegroup
#>     1:  [70,80)
#>     2:  [60,70)
#>     3:  [50,60)
#>     4:  [60,70)
#>     5:  [70,80)
#>    ---         
#> 11721:  [60,70)
#> 11722:  [60,70)
#> 11723:  [50,60)
#> 11724:  [50,60)
#> 11725:  [50,60)
# set hsize to 1,...,5+
eusilcS[, hsize := cut(hsize, c(0:4, Inf), labels = c(1:4, "5+"))]
#>        household hsize         state age gender netIncome    db090   weight
#>     1:         1     2      Salzburg  72   male  22675.48 7.822929 7.822929
#>     2:         1     2      Salzburg  66 female  16999.29 7.822929 7.822929
#>     3:         2     1 Upper Austria  56 female  19274.21 8.788089 8.788089
#>     4:         3     1        Styria  67 female  13319.13 8.108452 8.108452
#>     5:         4     3 Upper Austria  70 female  14365.57 7.509383 7.509383
#>    ---                                                                     
#> 11721:      4639     2     Carinthia  60 female  11405.85 6.863162 6.863162
#> 11722:      4640     2     Carinthia  60   male  15476.11 6.863162 6.863162
#> 11723:      4640     2     Carinthia  55 female      0.00 6.863162 6.863162
#> 11724:      4641     2        Vienna  56   male  30367.92 8.772773 8.772773
#> 11725:      4641     2        Vienna  52 female   6368.93 8.772773 8.772773
#>        agegroup
#>     1:  [70,80)
#>     2:  [60,70)
#>     3:  [50,60)
#>     4:  [60,70)
#>     5:  [70,80)
#>    ---         
#> 11721:  [60,70)
#> 11722:  [60,70)
#> 11723:  [50,60)
#> 11724:  [50,60)
#> 11725:  [50,60)
# treat households as a factor variable
eusilcS[, household := as.factor(household)]
#>        household hsize         state age gender netIncome    db090   weight
#>     1:         1     2      Salzburg  72   male  22675.48 7.822929 7.822929
#>     2:         1     2      Salzburg  66 female  16999.29 7.822929 7.822929
#>     3:         2     1 Upper Austria  56 female  19274.21 8.788089 8.788089
#>     4:         3     1        Styria  67 female  13319.13 8.108452 8.108452
#>     5:         4     3 Upper Austria  70 female  14365.57 7.509383 7.509383
#>    ---                                                                     
#> 11721:      4639     2     Carinthia  60 female  11405.85 6.863162 6.863162
#> 11722:      4640     2     Carinthia  60   male  15476.11 6.863162 6.863162
#> 11723:      4640     2     Carinthia  55 female      0.00 6.863162 6.863162
#> 11724:      4641     2        Vienna  56   male  30367.92 8.772773 8.772773
#> 11725:      4641     2        Vienna  52 female   6368.93 8.772773 8.772773
#>        agegroup
#>     1:  [70,80)
#>     2:  [60,70)
#>     3:  [50,60)
#>     4:  [60,70)
#>     5:  [70,80)
#>    ---         
#> 11721:  [60,70)
#> 11722:  [60,70)
#> 11723:  [50,60)
#> 11724:  [50,60)
#> 11725:  [50,60)

## example for base weights assuming a simple random sample of households stratified per region
eusilcS[, regSamp := .N, by = state]
#>        household hsize         state age gender netIncome    db090   weight
#>     1:         1     2      Salzburg  72   male  22675.48 7.822929 7.822929
#>     2:         1     2      Salzburg  66 female  16999.29 7.822929 7.822929
#>     3:         2     1 Upper Austria  56 female  19274.21 8.788089 8.788089
#>     4:         3     1        Styria  67 female  13319.13 8.108452 8.108452
#>     5:         4     3 Upper Austria  70 female  14365.57 7.509383 7.509383
#>    ---                                                                     
#> 11721:      4639     2     Carinthia  60 female  11405.85 6.863162 6.863162
#> 11722:      4640     2     Carinthia  60   male  15476.11 6.863162 6.863162
#> 11723:      4640     2     Carinthia  55 female      0.00 6.863162 6.863162
#> 11724:      4641     2        Vienna  56   male  30367.92 8.772773 8.772773
#> 11725:      4641     2        Vienna  52 female   6368.93 8.772773 8.772773
#>        agegroup regSamp
#>     1:  [70,80)     779
#>     2:  [60,70)     779
#>     3:  [50,60)    2120
#>     4:  [60,70)    1859
#>     5:  [70,80)    2120
#>    ---                 
#> 11721:  [60,70)     927
#> 11722:  [60,70)     927
#> 11723:  [50,60)     927
#> 11724:  [50,60)    1789
#> 11725:  [50,60)    1789
eusilcS[, regPop := sum(weight), by = state]
#>        household hsize         state age gender netIncome    db090   weight
#>     1:         1     2      Salzburg  72   male  22675.48 7.822929 7.822929
#>     2:         1     2      Salzburg  66 female  16999.29 7.822929 7.822929
#>     3:         2     1 Upper Austria  56 female  19274.21 8.788089 8.788089
#>     4:         3     1        Styria  67 female  13319.13 8.108452 8.108452
#>     5:         4     3 Upper Austria  70 female  14365.57 7.509383 7.509383
#>    ---                                                                     
#> 11721:      4639     2     Carinthia  60 female  11405.85 6.863162 6.863162
#> 11722:      4640     2     Carinthia  60   male  15476.11 6.863162 6.863162
#> 11723:      4640     2     Carinthia  55 female      0.00 6.863162 6.863162
#> 11724:      4641     2        Vienna  56   male  30367.92 8.772773 8.772773
#> 11725:      4641     2        Vienna  52 female   6368.93 8.772773 8.772773
#>        agegroup regSamp    regPop
#>     1:  [70,80)     779  5344.346
#>     2:  [60,70)     779  5344.346
#>     3:  [50,60)    2120 14212.777
#>     4:  [60,70)    1859 11668.709
#>     5:  [70,80)    2120 14212.777
#>    ---                           
#> 11721:  [60,70)     927  5637.013
#> 11722:  [60,70)     927  5637.013
#> 11723:  [50,60)     927  5637.013
#> 11724:  [50,60)    1789 16006.860
#> 11725:  [50,60)    1789 16006.860
eusilcS[, baseWeight := regPop/regSamp]
#>        household hsize         state age gender netIncome    db090   weight
#>     1:         1     2      Salzburg  72   male  22675.48 7.822929 7.822929
#>     2:         1     2      Salzburg  66 female  16999.29 7.822929 7.822929
#>     3:         2     1 Upper Austria  56 female  19274.21 8.788089 8.788089
#>     4:         3     1        Styria  67 female  13319.13 8.108452 8.108452
#>     5:         4     3 Upper Austria  70 female  14365.57 7.509383 7.509383
#>    ---                                                                     
#> 11721:      4639     2     Carinthia  60 female  11405.85 6.863162 6.863162
#> 11722:      4640     2     Carinthia  60   male  15476.11 6.863162 6.863162
#> 11723:      4640     2     Carinthia  55 female      0.00 6.863162 6.863162
#> 11724:      4641     2        Vienna  56   male  30367.92 8.772773 8.772773
#> 11725:      4641     2        Vienna  52 female   6368.93 8.772773 8.772773
#>        agegroup regSamp    regPop baseWeight
#>     1:  [70,80)     779  5344.346   6.860522
#>     2:  [60,70)     779  5344.346   6.860522
#>     3:  [50,60)    2120 14212.777   6.704140
#>     4:  [60,70)    1859 11668.709   6.276874
#>     5:  [70,80)    2120 14212.777   6.704140
#>    ---                                      
#> 11721:  [60,70)     927  5637.013   6.080920
#> 11722:  [60,70)     927  5637.013   6.080920
#> 11723:  [50,60)     927  5637.013   6.080920
#> 11724:  [50,60)    1789 16006.860   8.947379
#> 11725:  [50,60)    1789 16006.860   8.947379

## constraints on person level
# age 
conP1 <- xtabs(weight ~ agegroup, data = eusilcS)
# gender by region
conP2 <- xtabs(weight ~ gender + state, data = eusilcS)
# personal net income by gender
conP3 <- xtabs(weight*netIncome ~ gender, data = eusilcS)

## constraints on household level
conH1 <- xtabs(weight ~ hsize + state, data = eusilcS, subset = !duplicated(household))

# array of convergence limits for conH1
epsH1 <- conH1
epsH1[1:4,] <- 0.005
epsH1["5+",] <- 0.2

# without array epsH1
calibweights1 <- ipu2(eusilcS, hid = "household", 
                      conP = list(conP1, conP2, netIncome = conP3), 
                      conH = list(conH1), 
                      epsP = list(1e-06, 1e-06, 1e-03),
                      epsH = 0.01,  
                      bound = NULL, verbose = TRUE,  maxIter = 200)
#> Warning: the functionality of ipu2 moved to the package surveysd in the function ipf
#> 10:Not yet converged for P-Constraint1
#> 10:Not yet converged for P-Constraint2
#> 10:Not yet converged for P-Constraint3
#> 20:Not yet converged for P-Constraint1
#> 20:Not yet converged for P-Constraint2
#> 30:Not yet converged for P-Constraint1
#> 30:Not yet converged for P-Constraint2
#> 40:Not yet converged for P-Constraint1
#> 40:Not yet converged for P-Constraint2
#> Iteration stopped after 43 steps
#> Convergence reached

# with array epsH1, base weights and bound
calibweights2 <- ipu2(eusilcS, hid = "household", 
                      conP = list(conP1, conP2), 
                      conH = list(conH1), 
                      epsP = 1e-06,
                      epsH = list(epsH1),  
                      w = "baseWeight",
                      bound = 4, verbose = TRUE, maxIter = 200)
#> Warning: the functionality of ipu2 moved to the package surveysd in the function ipf
#> 10:Not yet converged for P-Constraint1
#> 10:Not yet converged for P-Constraint2
#> 10:Not yet converged for H-Constraint1
#> 20:Not yet converged for P-Constraint1
#> 20:Not yet converged for P-Constraint2
#> 30:Not yet converged for P-Constraint1
#> 30:Not yet converged for P-Constraint2
#> 40:Not yet converged for P-Constraint1
#> 40:Not yet converged for P-Constraint2
#> Iteration stopped after 46 steps
#> Convergence reached
                      
# show an adjusted version of conP and the original
attr(calibweights2, "conP_adj")
#> [[1]]
#> agegroup
#>  [-Inf,10)    [10,20)    [20,30)    [30,40)    [40,50)    [50,60)    [60,70) 
#>  7990.8360  9830.0050  9901.6674 12100.3417 13978.4222 10999.9179  8597.5028 
#>    [70,80)    [80,90)  [90, Inf) 
#>  5311.5742  2796.6133   313.1196 
#> 
#> [[2]]
#>         state
#> gender   Burgenland Carinthia Lower Austria Salzburg   Styria    Tyrol
#>   male     1193.554  2793.544      7817.991 2645.441 5627.668 3389.482
#>   female   1407.260  2843.469      7752.320 2698.905 6041.041 3623.173
#>         state
#> gender   Upper Austria   Vienna Vorarlberg
#>   male        6994.741 7569.677   1850.917
#>   female      7218.038 8437.181   1915.598
#> 
attr(calibweights2, "conP")
#> [[1]]
#> agegroup
#>  [-Inf,10)    [10,20)    [20,30)    [30,40)    [40,50)    [50,60)    [60,70) 
#>  7990.8432  9830.0105  9901.6665 12100.3345 13978.4171 10999.9180  8597.5030 
#>    [70,80)    [80,90)  [90, Inf) 
#>  5311.5743  2796.6133   313.1196 
#> 
#> [[2]]
#>         state
#> gender   Burgenland Carinthia Lower Austria Salzburg   Styria    Tyrol
#>   male     1193.555  2793.547      7817.994 2645.443 5627.671 3389.485
#>   female   1407.259  2843.466      7752.317 2698.904 6041.038 3623.170
#>         state
#> gender   Upper Austria   Vienna Vorarlberg
#>   male        6994.744 7569.680   1850.918
#>   female      7218.033 8437.181   1915.596
#>