Utility functions for socio-economic data sets

Various utility functions mainly used for simulating EU-SILC data

getAge(birth, year, data = NULL)

getGender(gender, labels = c("male", "female"), data = NULL)

getEcoStat(ecoStat, data, levels)

getCitizenship(citizenship, data, owncountry, EU, other)

getHsize(data, hhid)

restructureHHid(data, hhid)

factorNA(x, always = FALSE, newval = NA)

Arguments

birth

year of birth

year

current year

data

data.frame (for some functions optional)

gender

variable including information on gender

labels

labels of a factor variable

ecoStat

variable holding information on the economic status

hhid

name or index of variable holding the information on household ID

Author

Andreas Alfons, Matthias Templ

Examples

birth <- sample(1950:2000, 20)
getAge(birth, 2013)
#>  [1] 57 48 20 21 39 13 24 33 14 31 17 26 28 43 12 30 19 45 49 15
data(eusilcS)
head(getGender("rb090", labels = c("ma","fe"), data=eusilcS))
#> [1] ma fe fe fe fe ma
#> Levels: ma fe
lev <- c("Employee working full-time", "Employee working part-time",
         "Self-employed working full-time", "Self-employed working part-time",
         "Unemployed", "Pupil, student, further training, unpaid work experience",
         "In retirement", "Permanently disabled", "In compulsory military community or service",
         "Fulfilling domestic tasks", "Other inactive person")
g <- getEcoStat("pl030", eusilcS, lev)
table(g)
#> g
#>                               Employee working full-time 
#>                                                     4207 
#>                               Employee working part-time 
#>                                                      876 
#>                          Self-employed working full-time 
#>                                                      374 
#>                          Self-employed working part-time 
#>                                                      604 
#>                                               Unemployed 
#>                                                     2352 
#> Pupil, student, further training, unpaid work experience 
#>                                                      111 
#>                                            In retirement 
#>                                                      998 
#>                                     Permanently disabled 
#>                                                        0 
#>              In compulsory military community or service 
#>                                                        0 
#>                                Fulfilling domestic tasks 
#>                                                        0 
#>                                    Other inactive person 
#>                                                        0 
data(eusilcS)
## destroy info on pb220a to show afterwards the usage of the function
owncountry <- "AT"
EU <- c("DE","BE","BG","CY","CZ","DK","EE","EL","ES","FI","FR","GR","HU","IE",
        "IT","LT","LU","LV","MT","NL","PL","PT","RO","SI","SE","SK","UK")
other <- c("CAN","CH","CSA","HR","IS","ME","MK","NAF","NME","NO",
           "OAF","OAS","OCE","OEU","OT","OTH","TR","USA","WAF")
eusilcS$fakepb220a <-  factor(sample(c(owncountry, EU, other), nrow(eusilcS), replace = TRUE))
table(eusilcS$fakepb220a)
#> 
#>  AT  BE  BG CAN  CH CSA  CY  CZ  DE  DK  EE  EL  ES  FI  FR  GR  HR  HU  IE  IS 
#> 238 254 244 238 229 233 236 230 258 265 252 266 267 230 232 256 254 275 250 244 
#>  IT  LT  LU  LV  ME  MK  MT NAF  NL NME  NO OAF OAS OCE OEU  OT OTH  PL  PT  RO 
#> 263 234 255 251 262 255 243 239 271 252 263 266 265 235 234 248 242 224 259 265 
#>  SE  SI  SK  TR  UK USA WAF 
#> 258 244 240 285 264 233 224 
eusilcS$fakepb220a <- getCitizenship(citizenship = "fakepb220a",
                                  data=eusilcS, owncountry=owncountry,
                                  EU=EU, other=other)
table(eusilcS$fakepb220a)
#> 
#>    AT    EU Other 
#>   238  6786  4701 
data(eusilcS)
hsize <- getHsize(data=eusilcS)
table(hsize)
#> hsize
#>    1    2    3    4    5    6    7    8    9 
#> 1313 2770 2391 2752 1560  588  245   88   18 
hhid <- c(6,6,3,3,3,2,1,1,8,9,9,9,9,7,7)
hhid
#>  [1] 6 6 3 3 3 2 1 1 8 9 9 9 9 7 7
df <- data.frame("hhid"=hhid)
restructureHHid(df, "hhid")
#>  [1] 1 1 2 3 3 3 6 6 7 7 8 9 9 9 9
hhid <- factor(c(6,6,3,3,3,2,1,1,NA,9,9,9,9,7,7))
hhid
#>  [1] 6    6    3    3    3    2    1    1    <NA> 9    9    9    9    7    7   
#> Levels: 1 2 3 6 7 9
factorNA(hhid)
#>  [1] 6    6    3    3    3    2    1    1    <NA> 9    9    9    9    7    7   
#> Levels: 1 2 3 6 7 9 <NA>