Simulate categorical variables of population data

Simulate categorical variables of population data taking relationships between household members into account. The household structure of the population data needs to be simulated beforehand using simStructure().

simRelation(
  simPopObj,
  relation = "relate",
  head = "head",
  direct = NULL,
  additional,
  limit = NULL,
  censor = NULL,
  maxit = 500,
  MaxNWts = 2000,
  eps = NULL,
  nr_cpus = NULL,
  seed = 1,
  regModel = NULL,
  verbose = FALSE,
  method = c("multinom", "ctree", "cforest", "ranger"),
  by = "strata"
)

Arguments

simPopObj

a simPopObj containing population and household survey data as well as optionally margins in standardized format.

relation

a character string specifying the columns of dataS and dataP, respectively, that define the relationships between the household members.

head

a character string specifying the category of the variable given by relation that identifies the household head.

direct

a character string specifying categories of the variable given by relation. Simulated individuals with those categories directly inherit the values of the additional variables from the household head. The default is NULL such that no individuals directly inherit value from the household head.

additional

a character vector specifying additional categorical variables of dataS that should be simulated for the population data.

limit

this can be used to account for structural zeros. If only one additional variable is requested, a named list of lists should be supplied. The names of the list components specify the predictor variables for which to limit the possible outcomes of the response. For each predictor, a list containing the possible outcomes of the response for each category of the predictor can be supplied. The probabilities of other outcomes conditional on combinations that contain the specified categories of the supplied predictors are set to 0. If more than one additional variable is requested, such a list of lists can be supplied for each variable as a component of yet another list, with the component names specifying the respective variables.

censor

this can be used to account for structural zeros. If only one additional variable is requested, a named list of lists or data.frames should be supplied. The names of the list components specify the categories that should be censored. For each of these categories, a list or data.frame containing levels of the predictor variables can be supplied. The probability of the specified categories is set to 0 for the respective predictor levels. If more than one additional variable is requested, such a list of lists or data.frames can be supplied for each variable as a component of yet another list, with the component names specifying the respective variables.

maxit, MaxNWts

control parameters to be passed to nnet::multinom() and nnet::nnet(). See the help file for nnet::nnet().

eps

a small positive numeric value, or NULL (the default). In the former case, estimated probabilities smaller than this are assumed to result from structural zeros and are set to exactly 0.

nr_cpus

if specified, an integer number defining the number of cpus that should be used for parallel processing.

seed

optional; an integer value to be used as the seed of the random number generator, or an integer vector containing the state of the random number generator to be restored.

regModel

allows to specify the variables or model that is used when simulating additional categorical variables. The following choices are available if different from NULL.

"basic": only the basic household variables (generated with simStructure() are used.

"available": all available variables (that are common in the sample and the synthetic population such as previously generated variables) excluding id-variables, strata variables and household sizes are used for the modeling. This parameter should be used with care because all factors are automatically used as factors internally.

formula-object: users may also specify a formula (class 'formula') that will be used. Checks are performed that all required variables are available. If parameter regModel is NULL, only basic household variables are used in any case.

verbose

set to TRUE if additional print output should be shown.

method

a character string specifying the method to be used for simulating the additional categorical variables. Accepted values are

"multinom": estimation of the conditional probabilities using multinomial log-linear models and random draws from the resulting distributions

"ctree": for using Classification trees

"cforest": for using random forest (implementation in package party)

"ranger": for using random forest (implementation in package ranger)

by

defining which variable to use as split up variable of the estimation. Defaults to the strata variable.

Value

An object of class simPopObj containing survey data as well as the simulated population data including the categorical variables specified by additional.

Details

The values of a new variable are simulated in three steps, where the second step is optional. First, the values of the household heads are simulated with multinomial log-linear models. Second, individuals directly related to the corresponding household head (as specified by the argument direct) inherit the value of the latter. Third, the values of the remaining individuals are simulated with multinomial log-linear models in which the value of the respective household head is used as an additional predictor.

The number of cpus are selected automatically in the following manner. The number of cpus is equal the number of strata. However, if the number of cpus is less than the number of strata, the number of cpus - 1 is used by default. This should be the best strategy, but the user can also overwrite this decision.

Note

The basic household structure needs to be simulated beforehand with the function simStructure().

See also

Author

Andreas Alfons and Bernhard Meindl

Examples

data(ghanaS) # load sample data
samp <- specifyInput(
  data = ghanaS,
  hhid = "hhid",
  strata = "region",
  weight = "weight"
)
ghanaP <- simStructure(
  data = samp,
  method = "direct",
  basicHHvars = c("age", "sex", "relate")
)
class(ghanaP)
#> [1] "simPopObj"
#> attr(,"package")
#> [1] "simPop"

# \donttest{
## long computation time ...
ghanaP <- simRelation(
  simPopObj = ghanaP,
  relation = "relate",
  head = "head",
  additional = c("nation", "ethnic", "religion")
)
str(ghanaP)
#> Formal class 'simPopObj' [package "simPop"] with 4 slots
#>   ..@ sample     :Formal class 'dataObj' [package "simPop"] with 7 slots
#>   .. .. ..@ data        :Classes ‘data.table’ and 'data.frame':	36970 obs. of  16 variables:
#>   .. .. .. ..$ hhid          : int [1:36970] 1 1 1 2 2 2 2 3 4 4 ...
#>   .. .. .. ..$ hsize         : int [1:36970] 3 3 3 4 4 4 4 1 6 6 ...
#>   .. .. .. ..$ region        : Factor w/ 10 levels "western","central",..: 1 1 1 6 6 6 6 5 3 3 ...
#>   .. .. .. ..$ clust         : Factor w/ 580 levels "5001","5002",..: 14 14 14 297 297 297 297 265 157 157 ...
#>   .. .. .. ..$ age           : int [1:36970] 37 32 17 64 15 6 5 40 77 24 ...
#>   .. .. .. ..$ sex           : Factor w/ 2 levels "male","female": 1 2 1 2 1 1 2 1 2 1 ...
#>   .. .. .. ..$ relate        : Factor w/ 10 levels "head","spouse",..: 1 2 3 1 3 4 4 1 1 4 ...
#>   .. .. .. ..$ nation        : Factor w/ 11 levels "ghanaian birth",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. ..$ ethnic        : Factor w/ 9 levels "akan","all other tribes",..: 1 1 1 1 1 1 1 1 4 3 ...
#>   .. .. .. ..$ religion      : Factor w/ 11 levels "catholic","anglican",..: 5 5 7 7 1 7 1 6 3 1 ...
#>   .. .. .. ..$ highest_degree: Factor w/ 16 levels "none","mslc",..: 2 1 1 1 1 1 1 13 2 8 ...
#>   .. .. .. ..$ occupation    : Factor w/ 11 levels "armed forces and other security personnel",..: 4 6 6 10 6 6 6 8 6 4 ...
#>   .. .. .. ..$ income        : num [1:36970] 576.7 0 0 71.6 0 ...
#>   .. .. .. ..$ weight        : num [1:36970] 5.44 5.44 5.44 6.6 6.6 ...
#>   .. .. .. ..$ hhsize        : int [1:36970] 3 3 3 4 4 4 4 1 6 6 ...
#>   .. .. .. ..$ pid           : chr [1:36970] "1.1" "1.2" "1.3" "2.1" ...
#>   .. .. .. ..- attr(*, "sorted")= chr [1:2] "hhid" "pid"
#>   .. .. .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   .. .. ..@ hhid        : chr "hhid"
#>   .. .. ..@ pid         : chr "pid"
#>   .. .. ..@ hhsize      : chr "hhsize"
#>   .. .. ..@ weight      : chr "weight"
#>   .. .. ..@ strata      : chr "region"
#>   .. .. ..@ ispopulation: logi FALSE
#>   ..@ table      : NULL
#>   ..@ pop        :Formal class 'dataObj' [package "simPop"] with 7 slots
#>   .. .. ..@ data        :Classes ‘data.table’ and 'data.frame':	226451 obs. of  11 variables:
#>   .. .. .. ..$ hhid    : int [1:226451] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. .. ..$ hhsize  : int [1:226451] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. ..$ region  : Factor w/ 10 levels "western","central",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. ..$ age     : int [1:226451] 70 42 60 74 65 19 57 21 63 70 ...
#>   .. .. .. ..$ sex     : Factor w/ 2 levels "male","female": 2 1 2 1 2 1 2 1 1 1 ...
#>   .. .. .. ..$ relate  : Factor w/ 10 levels "head","spouse",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. ..$ pid     : chr [1:226451] "1.1" "2.1" "3.1" "4.1" ...
#>   .. .. .. ..$ weight  : num [1:226451] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. ..$ nation  : Factor w/ 11 levels "ghanaian birth",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. ..$ ethnic  : Factor w/ 9 levels "akan","all other tribes",..: 1 1 1 1 1 1 1 9 1 1 ...
#>   .. .. .. ..$ religion: Factor w/ 11 levels "catholic","anglican",..: 10 3 1 10 10 6 5 5 5 5 ...
#>   .. .. .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   .. .. .. ..- attr(*, "sorted")= chr "hhid"
#>   .. .. ..@ hhid        : chr "hhid"
#>   .. .. ..@ pid         : chr "pid"
#>   .. .. ..@ hhsize      : chr "hhsize"
#>   .. .. ..@ weight      : chr "weight"
#>   .. .. ..@ strata      : chr "region"
#>   .. .. ..@ ispopulation: logi TRUE
#>   ..@ basicHHvars: chr [1:3] "age" "sex" "relate"
# }