Simple generation of new variables

Fast simulation of new variables based on univariate distributions

univariate.dis(puf, data, additional, weights, value = "data", fNA = NA)

conditional.dis(
  puf,
  data,
  additional,
  conditional,
  weights,
  value = "data",
  fNA = NA
)

Arguments

puf

data for which one additional column specified by function argument ‘additional’ is simulated

data

donor data

additional

name of variable to be simulated

weights

sampling weights from data

value

if “data” then the puf including the additional variable is returned, otherwise only the simulated vector.

fNA

only used with missing values if another code as NA should be used

conditional

conditioning variable

Details

Function uni.distribution: random draws from the weighted univariate distribution of the original data

Function conditional.dis: random draws from the weighted conditional distribution (conditioned on a factor variable)

This are simple functions to produce structural variables, variables that should have the same categories as given ones. For more advanced methods see simCategorical()

See also

Author

Lydia Spies, Matthias Templ

Examples

## we don't have original data, so let's use eusilc
data(eusilc13puf)
data(eusilcS)
v1 <- univariate.dis(eusilcS, eusilc13puf, additional = "db040",
weights = "rb050", value = "vector")
table(v1)
#> v1
#> AT11 AT12 AT13 AT21 AT22 AT31 AT32 AT33 AT34 
#>  377 2269 2291  743 1699 2026  778 1000  542 
table(eusilc13puf$db040)
#> 
#> AT11 AT12 AT13 AT21 AT22 AT31 AT32 AT33 AT34 
#>  487 2691 2481  804 1883 2457  883 1220  607 
## we don't have original data, so let's use eusilc
##data(eusilc13puf)
##data(eusilcS)
##v1 <- conditional.dis(eusilcS, eusilc13puf, additional = "pb190",
##  conditional = "db040", weights = "rb050")
##table(v1) / sum(table(v1))
##table(eusilc13puf$pb190) / sum(table(eusilc13puf$pb190))