Weighted box plot statistics

Compute the statistics necessary for producing box-and-whisker plots of continuous or semi-continuous variables, taking into account sample weights.

spBwplotStats(x, weights = NULL, coef = 1.5, zeros = TRUE, do.out = TRUE)

Arguments

x

a numeric vector.

weights

an optional numeric vector containing sample weights.

coef

a numeric value that determines the extension of the whiskers.

zeros

a logical indicating whether the variable specified by additional is semi-continuous, i.e., contains a considerable amount of zeros. If TRUE, the (weighted) box plot statistics are computed for the non-zero data points only and the number of zeros is returned, too.

do.out

a logical indicating whether data points that lie beyond the extremes of the whiskers should be returned.

Value

A list of class "spBwplotStats" with the following components:

stats

A vector of length 5 containing the (weighted) statistics for the construction of a box plot.

n

if weights is NULL, the number of non-missing and, if zeros is TRUE, non-zero data points. Otherwise the sum of the weights of the corresponding points.

nzero

if zeros is TRUE and weights is NULL, the number of zeros. If zeros is TRUE and weights is not NULL, the sum of the weights of the zeros. If zeros is not TRUE, this is NULL.

out

if do.out, the values of any data points that lie beyond the extremes of the whiskers.

Details

The function quantileWt is used for the computation of (weighted) quantiles. The median is computed together with the first and the third quartile, which form the box. If range is positive, the whiskers extend to the most extreme data points that have a distance to the box of no more than coef times the interquartile range. For coef = 0, the whiskers mark the minimum and the maximum of the sample, whereas a negative value causes an error.

See also

Author

Stefan Kraft and Andreas Alfons

Examples

data(eusilcS)

## semi-continuous variable
spBwplotStats(eusilcS$netIncome, 
    weights=eusilcS$rb050, do.out = FALSE)
#> $stats
#> [1] -5809.40  9754.64 15812.21 22161.47 40757.02
#> 
#> $n
#> [1] 59920.31
#> 
#> $nzero
#> [1] 7882.757
#> 
#> $out
#> numeric(0)
#> 
#> attr(,"class")
#> [1] "spBwplotStats"