create an object of class 'dataObj' required for further processing

create an standardized input object of class 'dataObj' containing information on weights, household ids, household sizes, person ids and optionally strata. Outputs of this function are typically used in simStructure.

specifyInput(
  data,
  hhid,
  hhsize = NULL,
  pid = NULL,
  weight = NULL,
  strata = NULL,
  population = FALSE
)

Arguments

data

a data.frame or data.table featuring sample data.

hhid

character vector of length 1 specifying variable containing household ids within slot data.

hhsize

character vector of length 1 specifying variable containing household sizes within slot data. If NULL, household sizes are automatically calculated.

pid

character vector of length 1 specifying variable containing person ids within slot data. If NULL, person ids are automatically calculated.

weight

character vector of length 1 specifying variable holding sampling weights within slot data.

strata

character vector of length 1 specifing variable name within slot data of variable holding information on strata, e.g. regions or NULL if such variable does not exist.

population

TRUE/FALSE vector of length 1 specifing if the data object is a sample or a population NULL if such variable does not exist.

References

M. Templ, B. Meindl, A. Kowarik, A. Alfons, O. Dupriez (2017) Simulation of Synthetic Populations for Survey Data Considering Auxiliary Information. Journal of Statistical Survey, 79 (10), 1--38. doi:10.18637/jss.v079.i10

Author

Bernhard Meindl

Examples

data(eusilcS)
inp <- specifyInput(data=eusilcS, hhid="db030", weight="rb050", strata="db040")
class(inp)
#> [1] "dataObj"
#> attr(,"package")
#> [1] "simPop"
inp
#> 
#>  -------------- 
#> survey sample of size 11725 x 20 
#> 
#>  Selected important variables: 
#> 
#>  household ID: db030
#>  personal ID: pid
#>  variable household size: hhsize
#>  sampling weight: rb050
#>  strata: db040
#>  -------------- 
#>