Sprague index (multipliers)

Using the Sprague multipliers, the age counts are estimated for each year having 5-years interval data as input.

sprague(x)

Arguments

x

numeric vector of age counts in five-year intervals

Value

Population counts for age 0, 1, 2, 3, 4, ..., 78, 79, 80+.

Details

The input is population counts of age classes 0-4, 5-9, 10-14, ... , 77-74, 75-79, 80+.

References

G. Calot and J.-P. Sardon. Methodology for the calculation of Eurostat's demographic indicators. Detailed report by the European Demographic Observatory

See also

Author

Matthias Templ

Examples

## example from the world bank
x <- data.frame(age=as.factor(c(
  "0-4",
  "5-9","10-14","15-19", "20-24",
  "25-29","30-34","35-39","40-44","45-49",
  "50-54","55-59","60-64","65-69","77-74","75-79","80+"
    )),
  pop=c(1971990, 2095820,2157190, 2094110,2116580,   2003840, 1785690,
        1502990, 1214170, 796934,  627551,  530305, 488014,
        364498, 259029,158047,  125941)
)

s  <- sprague(x[,2])
s
#>         0         1         2         3         4         5         6         7 
#> 383576.38 388941.36 394401.76 399858.40 405212.10 410363.66 415213.92 419663.68 
#>         8         9        10        11        12        13        14        15 
#> 423613.76 426964.98 430053.33 433214.82 434174.34 431962.34 427785.18 423465.58 
#>        16        17        18        19        20        21        22        23 
#> 418215.14 415380.74 416676.82 420371.73 423067.07 425499.74 426001.26 423416.62 
#>        24        25        26        27        28        29        30        31 
#> 418595.30 413607.14 408116.62 401682.06 394315.74 386118.43 377272.10 367852.53 
#>        32        33        34        35        36        37        38        39 
#> 357766.85 347024.21 335774.32 323962.07 311363.67 299490.31 289013.80 279160.14 
#>        40        41        42        43        44        45        46        47 
#> 269085.55 259681.23 247052.14 229407.21 208943.86 189369.48 169605.22 153871.27 
#>        48        49        50        51        52        53        54        55 
#> 144536.03 139552.00 134475.80 130163.12 125947.72 121041.15 115923.21 111644.30 
#>        56        57        58        59        60        61        62        63 
#> 107732.88 104859.77 103378.29 102689.76 102001.01 101661.29  99759.49  95324.28 
#>        64        65        66        67        68        69        70        71 
#>  89267.93  83452.36  77479.16  72033.10  67634.07  63899.31  60029.19  56307.27 
#>        72        73        74        75        76        77        78        79 
#>  52268.78  47642.59  42781.17  38411.13  34510.77  31058.39  28157.02  25909.69 
#>       80+ 
#> 125941.00 
  
all.equal(sum(s), sum(x[,2]))
#> [1] TRUE