Weighted cross tabulation

Compute contingency tables taking into account sample weights.

tableWt(x, weights = NULL, useNA = c("no", "ifany", "always"))

Arguments

x

a vector that can be interpreted as a factor, or a matrix or data.frame whose columns can be interpreted as factors.

weights

an optional numeric vector containing sample weights.

useNA

a logical indicating whether to include extra NA levels in the table.

Value

The (weighted) contingency table as an object of class table, an array of integer values.

Details

For each combination of the variables in x, the weighted number of occurence is computed as the sum of the corresponding sample weights. If weights are not specified, the function table is applied.

See also

Author

Andreas Alfons and Stefan Kraft

Examples

data(eusilcS)
tableWt(eusilcS[, c("hsize", "db040")], weights = eusilcS$rb050)
#>      db040
#> hsize Burgenland Carinthia Lower Austria Salzburg Styria Tyrol Upper Austria
#>     1        302       737          2121      713   1630   808          1696
#>     2        787      1359          3568     1252   2796  1687          3172
#>     3        527      1264          3276      914   2690  1194          3086
#>     4        657      1304          3754     1510   2374  2009          3306
#>     5        250       504          2202      540   1114   878          1874
#>     6          0       338           358      229    687   234           603
#>     7         77        92           257      133    179    39           331
#>     8          0        39            35       53    154   165           108
#>     9          0         0             0        0     45     0            36
#>      db040
#> hsize Vienna Vorarlberg
#>     1   3770        416
#>     2   4632        745
#>     3   3366        702
#>     4   2556       1028
#>     5   1110        678
#>     6    348        173
#>     7    114         25
#>     8    110          0
#>     9      0          0
tableWt(eusilcS[, c("rb090", "pb220a")], weights = eusilcS$rb050, 
    useNA = "ifany")
#>         pb220a
#> rb090       AT    EU Other  <NA>
#>   male   29646   728  2258  7251
#>   female 32133   868  2170  6766