Population totals Region times Gender for Austria 2006

Population characteristics Region times Gender from Austria.

Using samp samp<- it is possible to extract or rather modify variables of the sample data within slot data in slot sample of the simPopObj-class-object. Using pop pop<- it is possible to extract or rather modify variables of the synthetic population within in slot data in slot sample of the simPopObj-class-object.

Format

totalsRG: A data frame with 18 observations on the following 3 variables.

list("rb090")

gender; a factor with levels female male

list("db040")

region; a factor with levels Burgenland Carinthia Lower Austria, Salzburg Styria Tyrol Upper Austria Vienna Vorarlberg

list("Freq")

totals; a numeric vector

totalsRGtab: a two-dimensional table holding the same information

totalsRG: A data frame with 18 observations on the following 3 variables.

list("rb090")

gender; a factor with levels female male

list("db040")

region; a factor with levels Burgenland Carinthia Lower Austria, Salzburg Styria Tyrol Upper Austria Vienna Vorarlberg

list("Freq")

totals; a numeric vector

totalsRGtab: a two-dimensional table holding the same information

Source

StatCube - statistical data base, http://www.statistik.at

StatCube - statistical data base, http://www.statistik.at/

Details

Population totals Region times Gender for Austria 2006

Population characteristics Region times Gender from Austria.

Examples

data(totalsRG)
totalsRG
#>     rb090         db040   Freq
#> 1    male    Burgenland 140436
#> 2  female    Burgenland 146980
#> 3    male     Carinthia 270084
#> 4  female     Carinthia 285797
#> 5    male Lower Austria 797398
#> 6  female Lower Austria 828087
#> 7    male      Salzburg 702539
#> 8  female      Salzburg 722883
#> 9    male        Styria 259595
#> 10 female        Styria 274675
#> 11   male         Tyrol 595842
#> 12 female         Tyrol 619404
#> 13   male Upper Austria 353910
#> 14 female Upper Austria 368128
#> 15   male        Vienna 850596
#> 16 female        Vienna 916150
#> 17   male    Vorarlberg 184939
#> 18 female    Vorarlberg 190343
data(totalsRGtab)
totalsRGtab
#>         db040
#> rb090    Burgenland Carinthia Lower Austria Salzburg Styria  Tyrol
#>   female     146980    285797        828087   722883 274675 619404
#>   male       140436    270084        797398   702539 259595 595842
#>         db040
#> rb090    Upper Austria Vienna Vorarlberg
#>   female        368128 916150     190343
#>   male          353910 850596     184939
data(totalsRG)
totalsRG
#>     rb090         db040   Freq
#> 1    male    Burgenland 140436
#> 2  female    Burgenland 146980
#> 3    male     Carinthia 270084
#> 4  female     Carinthia 285797
#> 5    male Lower Austria 797398
#> 6  female Lower Austria 828087
#> 7    male      Salzburg 702539
#> 8  female      Salzburg 722883
#> 9    male        Styria 259595
#> 10 female        Styria 274675
#> 11   male         Tyrol 595842
#> 12 female         Tyrol 619404
#> 13   male Upper Austria 353910
#> 14 female Upper Austria 368128
#> 15   male        Vienna 850596
#> 16 female        Vienna 916150
#> 17   male    Vorarlberg 184939
#> 18 female    Vorarlberg 190343
data(totalsRGtab)
totalsRGtab
#>         db040
#> rb090    Burgenland Carinthia Lower Austria Salzburg Styria  Tyrol
#>   female     146980    285797        828087   722883 274675 619404
#>   male       140436    270084        797398   702539 259595 595842
#>         db040
#> rb090    Upper Austria Vienna Vorarlberg
#>   female        368128 916150     190343
#>   male          353910 850596     184939