add known margins/totals

add known margins/totals for a combination of variables for the population to an object of class simPopObj.

addKnownMargins(inp, margins)

Arguments

inp

a simPopObj containing population and household survey data as well as optionally margins in standardized format.

margins

a data.frame containing for a combination of unique variable levels for n-variables the number of known occurences in the population. The numbers must be listed in the last column of data.frame 'margins' while the characteristics must be listed in the first 'n' columns.

Value

an object of class simPopObj with updated slot 'table'.

Details

The function takes a data.frame containing known marginals/totals for a some variables that must exist in the population (stored in slot 'pop' of input object 'inp') and updates slot 'table' of the input object. This slot finally contains the known totals.

households are drawn from the data and new ID's are generated for the new households.

References

M. Templ, B. Meindl, A. Kowarik, A. Alfons, O. Dupriez (2017) Simulation of Synthetic Populations for Survey Data Considering Auxiliary Information. Journal of Statistical Survey, 79 (10), 1--38. doi: 10.18637/jss.v079.i10

Author

Bernhard Meindl

Examples

data(eusilcS)
data(eusilcP)
# \donttest{
## approx. 20 seconds computation time
inp <- specifyInput(data=eusilcS, hhid="db030", hhsize="hsize", strata="db040", weight="db090")
inp <- simStructure(data=inp, method="direct", basicHHvars=c("age", "rb090"))
inp <- simCategorical(inp, additional=c("pl030", "pb220a"), method="multinom",nr_cpus=1)

margins <- as.data.frame(
  xtabs(rep(1, nrow(eusilcP)) ~ eusilcP$region + eusilcP$gender + eusilcP$citizenship))
colnames(margins) <- c("db040", "rb090", "pb220a", "freq")
inp <- addKnownMargins(inp, margins)
str(inp)
#> Formal class 'simPopObj' [package "simPop"] with 4 slots
#>   ..@ sample     :Formal class 'dataObj' [package "simPop"] with 7 slots
#>   .. .. ..@ data        :Classes ‘data.table’ and 'data.frame':	11725 obs. of  19 variables:
#>   .. .. .. ..$ db030    : int [1:11725] 1 1 2 3 4 4 4 5 5 5 ...
#>   .. .. .. ..$ hsize    : int [1:11725] 2 2 1 1 3 3 3 5 5 5 ...
#>   .. .. .. ..$ db040    : Factor w/ 9 levels "Burgenland","Carinthia",..: 4 4 7 5 7 7 7 4 4 4 ...
#>   .. .. .. ..$ age      : int [1:11725] 72 66 56 67 70 46 37 41 35 9 ...
#>   .. .. .. ..$ rb090    : Factor w/ 2 levels "male","female": 1 2 2 2 2 1 1 1 2 2 ...
#>   .. .. .. ..$ pl030    : Factor w/ 7 levels "1","2","3","4",..: 5 5 2 5 5 3 1 1 3 NA ...
#>   .. .. .. ..$ pb220a   : Factor w/ 3 levels "AT","EU","Other": 1 1 1 1 1 1 3 1 1 NA ...
#>   .. .. .. ..$ netIncome: num [1:11725] 22675 16999 19274 13319 14366 ...
#>   .. .. .. ..$ py010n   : num [1:11725] 0 0 19274 0 0 ...
#>   .. .. .. ..$ py050n   : num [1:11725] 0 0 0 0 0 ...
#>   .. .. .. ..$ py090n   : num [1:11725] 0 0 0 0 0 ...
#>   .. .. .. ..$ py100n   : num [1:11725] 22675 0 0 13319 14366 ...
#>   .. .. .. ..$ py110n   : num [1:11725] 0 0 0 0 0 0 0 0 0 NA ...
#>   .. .. .. ..$ py120n   : num [1:11725] 0 0 0 0 0 0 0 0 0 NA ...
#>   .. .. .. ..$ py130n   : num [1:11725] 0 16999 0 0 0 ...
#>   .. .. .. ..$ py140n   : num [1:11725] 0 0 0 0 0 0 0 0 0 NA ...
#>   .. .. .. ..$ db090    : num [1:11725] 7.82 7.82 8.79 8.11 7.51 ...
#>   .. .. .. ..$ rb050    : num [1:11725] 7.82 7.82 8.79 8.11 7.51 ...
#>   .. .. .. ..$ pid      : chr [1:11725] "1.1" "1.2" "2.1" "3.1" ...
#>   .. .. .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   .. .. .. ..- attr(*, "sorted")= chr [1:2] "db030" "pid"
#>   .. .. ..@ hhid        : chr "db030"
#>   .. .. ..@ pid         : chr "pid"
#>   .. .. ..@ hhsize      : chr "hsize"
#>   .. .. ..@ weight      : chr "db090"
#>   .. .. ..@ strata      : chr "db040"
#>   .. .. ..@ ispopulation: logi FALSE
#>   ..@ table      :Classes ‘data.table’ and 'data.frame':	54 obs. of  4 variables:
#>   .. ..$ db040 : chr [1:54] "Burgenland" "Burgenland" "Burgenland" "Burgenland" ...
#>   .. ..$ rb090 : chr [1:54] "male" "male" "male" "female" ...
#>   .. ..$ pb220a: chr [1:54] "AT" "EU" "Other" "AT" ...
#>   .. ..$ N     : num [1:54] 797 26 10 821 29 ...
#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..@ pop        :Formal class 'dataObj' [package "simPop"] with 7 slots
#>   .. .. ..@ data        :Classes ‘data.table’ and 'data.frame':	81838 obs. of  9 variables:
#>   .. .. .. ..$ db030 : int [1:81838] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. .. .. ..$ hsize : int [1:81838] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. ..$ db040 : Factor w/ 9 levels "Burgenland","Carinthia",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. ..$ age   : int [1:81838] 53 70 82 37 74 78 43 70 75 74 ...
#>   .. .. .. ..$ rb090 : Factor w/ 2 levels "male","female": 2 1 2 1 2 2 1 1 1 2 ...
#>   .. .. .. ..$ pid   : chr [1:81838] "1.1" "2.1" "3.1" "4.1" ...
#>   .. .. .. ..$ weight: num [1:81838] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. ..$ pl030 : Factor w/ 7 levels "1","2","3","4",..: 2 5 5 1 5 5 1 5 5 5 ...
#>   .. .. .. ..$ pb220a: Factor w/ 3 levels "AT","EU","Other": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   .. .. .. ..- attr(*, "sorted")= chr "db030"
#>   .. .. ..@ hhid        : chr "db030"
#>   .. .. ..@ pid         : chr "pid"
#>   .. .. ..@ hhsize      : chr "hsize"
#>   .. .. ..@ weight      : chr "weight"
#>   .. .. ..@ strata      : chr "db040"
#>   .. .. ..@ ispopulation: logi TRUE
#>   ..@ basicHHvars: chr [1:2] "age" "rb090"
# }