Cross tabulations of expected and realized population sizes.

Compute contingency tables of expected (i.e., estimated) and realized (i.e., simulated) population sizes. The expected values are obtained with the Horvitz-Thompson estimator.

spTable(inp, select)

Arguments

inp

an object of class simPopObj containing household survey and simulated population data.

select

character; vector defining the columns in slots 'pop' and 'sample' of argument 'input' that should be used for tabulation.

Value

A list of class "spTable" with the following components:

expected

the contingency table estimated from the survey data.

realized

the contingency table computed from the simulated population data.

Details

The contingency tables are computed with tableWt.

Note

Sampling weights are automatically used from the input object 'inp'!

See also

Author

Andreas Alfons and Bernhard Meindl

Examples

set.seed(1234)  # for reproducibility
data(eusilcS)   # load sample data
samp <- specifyInput(data=eusilcS, hhid="db030", hhsize="hsize",
  strata="db040", weight="db090")
eusilcP <- simStructure(data=samp, method="direct", basicHHvars=c("age", "rb090"))
res <- spTable(eusilcP, select = c("age", "rb090"))
class(res)
#> [1] "spTable"
res
#> $expected
#>     rb090
#> age  male female
#>   -1  166     70
#>   0   416    350
#>   1   394    336
#>   2   367    379
#>   3   372    423
#>   4   397    356
#>   5   379    350
#>   6   376    394
#>   7   509    431
#>   8   313    348
#>   9   488    374
#>   10  528    406
#>   11  477    535
#>   12  495    466
#>   13  533    490
#>   14  419    547
#>   15  621    511
#>   16  478    491
#>   17  534    462
#>   18  469    394
#>   19  450    527
#>   20  510    601
#>   21  648    453
#>   22  452    471
#>   23  502    601
#>   24  709    461
#>   25  510    582
#>   26  361    351
#>   27  386    462
#>   28  405    446
#>   29  553    438
#>   30  438    652
#>   31  531    613
#>   32  618    554
#>   33  565    683
#>   34  608    656
#>   35  666    620
#>   36  644    515
#>   37  648    645
#>   38  538    597
#>   39  654    655
#>   40  534    733
#>   41  808    850
#>   42  657    741
#>   43  706    917
#>   44  731    581
#>   45  806    719
#>   46  714    539
#>   47  659    689
#>   48  762    589
#>   49  629    614
#>   50  585    661
#>   51  521    476
#>   52  564    619
#>   53  491    595
#>   54  481    530
#>   55  435    744
#>   56  607    629
#>   57  644    460
#>   58  511    505
#>   59  428    513
#>   60  374    395
#>   61  445    627
#>   62  475    435
#>   63  436    426
#>   64  549    414
#>   65  443    453
#>   66  549    393
#>   67  318    517
#>   68  379    293
#>   69  230    446
#>   70  260    230
#>   71  273    392
#>   72  287    301
#>   73  286    337
#>   74  218    232
#>   75  337    321
#>   76  243    168
#>   77  195    369
#>   78  165    229
#>   79  202    267
#>   80  113    379
#>   81   89    295
#>   82  142    291
#>   83   93    279
#>   84   66    177
#>   85   72    204
#>   86   33    206
#>   87   38    103
#>   88   60     73
#>   89    0     86
#>   90   29     45
#>   91    9     97
#>   92   44     31
#>   93    0     15
#>   94    9      0
#>   95    8      4
#>   96   13      9
#> 
#> $realized
#>     rb090
#> age  male female
#>   -1  164     88
#>   0   413    360
#>   1   373    323
#>   2   358    369
#>   3   371    422
#>   4   392    379
#>   5   396    328
#>   6   379    426
#>   7   501    436
#>   8   309    363
#>   9   472    364
#>   10  511    425
#>   11  479    542
#>   12  482    470
#>   13  544    457
#>   14  426    579
#>   15  605    498
#>   16  470    460
#>   17  520    468
#>   18  475    409
#>   19  452    560
#>   20  492    592
#>   21  636    440
#>   22  411    455
#>   23  516    605
#>   24  731    466
#>   25  552    593
#>   26  371    355
#>   27  398    458
#>   28  425    472
#>   29  524    450
#>   30  427    668
#>   31  593    607
#>   32  592    553
#>   33  571    718
#>   34  593    617
#>   35  625    629
#>   36  632    519
#>   37  658    629
#>   38  514    569
#>   39  636    658
#>   40  529    695
#>   41  775    826
#>   42  666    751
#>   43  712    905
#>   44  746    613
#>   45  823    694
#>   46  730    564
#>   47  680    694
#>   48  741    597
#>   49  642    679
#>   50  582    595
#>   51  513    488
#>   52  578    637
#>   53  487    613
#>   54  491    521
#>   55  463    741
#>   56  634    657
#>   57  624    444
#>   58  482    474
#>   59  421    495
#>   60  375    390
#>   61  431    625
#>   62  473    439
#>   63  404    429
#>   64  573    404
#>   65  458    440
#>   66  601    395
#>   67  323    524
#>   68  364    281
#>   69  203    440
#>   70  248    243
#>   71  279    391
#>   72  286    298
#>   73  292    340
#>   74  209    233
#>   75  338    331
#>   76  236    189
#>   77  178    355
#>   78  167    239
#>   79  210    279
#>   80  131    378
#>   81  102    295
#>   82  142    297
#>   83  109    288
#>   84   57    176
#>   85   74    230
#>   86   33    203
#>   87   44     88
#>   88   51     79
#>   89    0     94
#>   90   30     45
#>   91   10     86
#>   92   40     15
#>   93    0     14
#>   94    7      0
#>   95   10      6
#>   96   14      7
#> 
#> attr(,"class")
#> [1] "spTable"